Move generic GenerateRoutes function to provider pkg
